+++ packages/soda-antd/CHANGELOG.md
@@ -1,5 +1,14 @@
 # soda-antd
 
+## 0.12.3
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.43.0
+    - soda-hooks@6.15.3
+    - soda-react@0.1.3
+
 ## 0.12.2
 
 ### Patch Changes